Return IList when we can.

+++ Source/RhythmCodex.Lib/Ssq/Model/Parameter0.cs
@@ -5,6 +5,6 @@
         public static short Timings => 0x0001;
         public static short Triggers => 0x0002;
         public static short Steps => 0x0003;
-        public static int Meta => 0x0009;
+        public static short Meta => 0x0009;
     }
 }
